Move plateFunctions
