+++ ui/src/KaptureApp/store/index.js
@@ -1,5 +1,3 @@
-import * as plateFunctions from './plateFunctions';
 export { store } from './store';
 export { editorComponentsActions } from './editorComponents';
 export { editorToolsActions } from './editorTools';
-export { plateFunctions };
