+++ ui/src/KaptureApp/utils/plateFunctions.js
